// src/interfaces/repo.ts
export interface Repo {
    owner:         string;
    repo:          string;
    link:          string;
    description?:  string;
    image:         string;
    website?:      string;
    language:      string;
    languageColor: string;
    stars:         number | string;
    forks:         number;
}